+++ attributes/default.rb
@@ -1,10 +1,11 @@
 default['logstash']['basedir'] = '/opt/logstash'
 default['logstash']['user'] = 'logstash'
 default['logstash']['group'] = 'logstash'
-default['logstash']['supervisor_gid'] = '' # set this if you want all logstash processes to run as different gid,  can be overridden
+default['logstash']['supervisor_gid'] = node['logstash']['group']
 default['logstash']['pid_dir'] = '/var/run/logstash'
 default['logstash']['create_account'] = true
 default['logstash']['join_groups'] = []
+
 # roles/flags for various search/discovery
 default['logstash']['graphite_role'] = 'graphite_server'
 default['logstash']['graphite_query'] = "roles:#{node['logstash']['graphite_role']} AND chef_environment:#{node.chef_environment}"
@@ -17,6 +18,7 @@ default['logstash']['graphite_ip'] = ''
 
 default['logstash']['patterns'] = {}
 default['logstash']['install_zeromq'] = false
+default['logstash']['install_rabbitmq'] = false
 
 case node['platform_family']
 when "rhel"
@@ -24,3 +26,12 @@ when "rhel"
 when "debian"
   default['logstash']['zeromq_packages'] = [ "zeromq",  "libzmq-dev"]
 end
+
+
+# Logging features
+default['logstash']['logging']['rotateFrequency'] = "daily"
+default['logstash']['logging']['maxBackup'] = 10
+default['logstash']['logging']['maxSize'] = "10M"
+default['logstash']['logging']['useFileSize'] = false
+
+