+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 This file is used to list changes made in each version of chef-logstash.
 
+## 0.7.4:
+* bump logstash version to 1.3.1
+
+## 0.7.3:
+* support for sudo with upstart for agent and server
+
+## 0.7.2:
+* embedded kibana support
+
+## 0.7.1:
+* various bugfixes
+* support for multiple logstash workers via attribute.
+
 ## 0.7.0:
 
 ### New features ###